+++ tc/autotuner/genetic_tuning_harness.cc
@@ -146,7 +146,7 @@ std::vector<size_t> inputDivisorsAndPowers2(
 }
 
 size_t largestDim(const std::vector<const DLTensor*>& inputs) {
-  CHECK_GE(inputs.size(), 0);
+  CHECK_GE(inputs.size(), 0u);
   auto maxElement = std::max_element(
       inputs.begin(), inputs.end(), [](const DLTensor* a, const DLTensor* b) {
         return a->ndim < b->ndim;
@@ -157,7 +157,7 @@ size_t largestDim(const std::vector<const DLTensor*>& inputs) {
 } // namespace
 
 void GeneticTunerHarness::setupTuningParameters() {
-  CHECK_GT(kInputs_.size(), 0);
+  CHECK_GT(kInputs_.size(), 0u);
   auto range = inputDivisorsAndPowers2(kInputs_.begin()->second);
   auto rangeUpTo64 = filterHigherThan(range, 64);
 
@@ -347,9 +347,9 @@ void GeneticTunerHarness::doGpuWork(
     ExecutorType& engine,
     Printer& printer) {
   WithDevice wd(gpu);
-  CHECK_EQ(1, kInputs_.count(gpu));
+  CHECK_EQ(1u, kInputs_.count(gpu));
   auto& inputs = kInputs_.at(gpu);
-  CHECK_EQ(1, outputs_.count(gpu));
+  CHECK_EQ(1u, outputs_.count(gpu));
   auto& outputs = outputs_.at(gpu);
 
   while (true) {
@@ -485,7 +485,7 @@ void GeneticTunerHarness::runOneGeneration(size_t generation) {
     currentCompilationJob_.store(0);
     numEvaluations_.store(0);
     readyToEvaluate_.resize(0);
-    for (int i = 0; i < kMaxPopulationSize; ++i) {
+    for (size_t i = 0; i < kMaxPopulationSize; ++i) {
       readyToEvaluate_.emplace_back();
       readyToEvaluate_[i].store(false);
     }
@@ -510,7 +510,7 @@ void GeneticTunerHarness::runOneGeneration(size_t generation) {
         cpuCompilationThread.join();
       }
     });
-    for (int i = 0; i < FLAGS_tuner_threads; ++i) {
+    for (size_t i = 0; i < FLAGS_tuner_threads; ++i) {
       cpuCompilationThreads.emplace_back(
           [this, &engine]() { this->doCompile(engine); });
     }
