+++ tc/core/mapping_options-inl.h
@@ -41,14 +41,14 @@ size_t TilingView::size() const {
 }
 
 ValueAccessor<uint64_t> TilingView::operator[](size_t i) {
-  CHECK_LT(i, proto.sizes_size()) << "index overflow";
+  CHECK_LT(i, static_cast<size_t>(proto.sizes_size())) << "index overflow";
   return ValueAccessor<uint64_t>(
       [this, i](uint64_t u) { this->proto.set_sizes(i, u); },
       [this, i]() { return this->proto.sizes(i); });
 }
 
 uint64_t TilingView::operator[](size_t i) const {
-  CHECK_LT(i, proto.sizes_size()) << "index overflow";
+  CHECK_LT(i, static_cast<size_t>(proto.sizes_size())) << "index overflow";
   return proto.sizes(i);
 }
